packages/cdk/pyproject.toml
[project]
name = "da-vinci-cdk"
version = "3.0.0"
description = "Da Vinci CDK components"
authors = [{ name = "Jim Rosser" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.184.1",
    "constructs>=10.0.0",
    "da-vinci==3.0.0",
]

[tool.uv.sources]
da-vinci = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["da_vinci_cdk"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
